+++ libs/auth/src/index.ts
@@ -1,2 +1,4 @@
 export * from './lib/models/auth-user';
 export * from './lib/models/auth-service';
+export * from './lib/tokens/auth-service.token';
+export * from './lib/guards/auth-guard';